requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
supabase==2.11.0
pydantic-settings==2.7.1
bcrypt==4.2.1
